+++ src/nti/externalization/zcml.py
@@ -11,16 +11,18 @@ from __future__ import division
 from __future__ import print_function
 
 from ZODB import loglevels
-
 from zope import interface
 from zope.component import zcml as component_zcml
 from zope.component.factory import Factory
+from zope.configuration.fields import Bool
 from zope.configuration.fields import GlobalInterface
 from zope.configuration.fields import GlobalObject
 from zope.configuration.fields import Tokens
 
+from nti.externalization import internalization
 from nti.externalization.autopackage import AutoPackageSearchingScopedInterfaceObjectIO
 from nti.externalization.interfaces import IMimeObjectFactory
+from nti.externalization.internalization import _find_factories_in_module
 
 __docformat__ = "restructuredtext en"
 
@@ -71,18 +73,8 @@ def registerMimeFactories(_context, module):
 
     :param module module: The module to inspect.
     """
-    # This is a pretty loose check. We can probably do better. For example,
-    # pass an interface parameter and only register things that provide
-    # that interface, or at least check to see if they are a ``type``
-    mod_name = module.__name__
-    for object_name, value in vars(module).items():
+    for object_name, value in _find_factories_in_module(module, case_sensitive=True):
         __traceback_info__ = object_name, value
-
-        try:
-            ext_create = value.__external_can_create__
-            v_mod_name = value.__module__
-        except AttributeError:
-            continue
 
         try:
             mime_type = value.mimeType
@@ -92,7 +84,7 @@ def registerMimeFactories(_context, module):
             except AttributeError:
                 continue
 
-        if mime_type and ext_create and mod_name == v_mod_name:
+        if mime_type:
             logger.log(loglevels.TRACE,
                        "Registered mime factory utility %s = %s (%s)",
                        object_name, value, mime_type)
@@ -124,14 +116,16 @@ class IAutoPackageExternalizationDirective(interface.Interface):
         title=u"The root interfaces defined by the package.",
         value_type=GlobalInterface(),
         required=True)
+
     modules = Tokens(
         title=u"Module names that contain the implementations of the root_interfaces.",
         value_type=GlobalObject(),
         required=True)
 
     factory_modules = Tokens(
-        title=u"If given, module names that should be searched for internalization factories",
-        description=u"If not given, all modules will be examined.",
+        title=u"If given, module names that should be searched for internalization factories.",
+        description=(u"If not given, all *modules* will be examined. If given, "
+                     u"**only** these modules will be searched."),
         value_type=GlobalObject(),
         required=False)
 
@@ -140,11 +134,31 @@ class IAutoPackageExternalizationDirective(interface.Interface):
                u"You can customize aspects of externalization that way."),
         required=False)
 
+    register_legacy_search_module = Bool(
+        title=(u"Register found factories by their class name."),
+        description=(u"If true (*not* the default), then, in addition to registering "
+                     "factories by their mime type, also register them all by their class name. "
+                     "This is not recommended; currently no conflicts are caught and the order "
+                     "is ill-defined. "
+                     "See https://github.com/NextThought/nti.externalization/issues/33"),
+        default=False,
+        required=False,
+    )
+
 
 
 
 def autoPackageExternalization(_context, root_interfaces, modules,
-                               factory_modules=None, iobase=None):
+                               factory_modules=None, iobase=None,
+                               register_legacy_search_module=False):
+    """
+    Implement the :class:`IAutoPackageExternalizationDirective` directive.
+
+    .. versionchanged:: 1.0
+       Add the *register_legacy_search_module* keyword argument, defaulting to
+       False. Previously legacy search modules would always be registered, but
+       now you must explicitly ask for it.
+    """
 
     ext_module_name = root_interfaces[0].__module__
     package_name = ext_module_name.rsplit('.', 1)[0]
@@ -189,17 +203,31 @@ def autoPackageExternalization(_context, root_interfaces, modules,
 
     # Now init the class so that it can add the things that internalization
     # needs.
-    # FIXME: We are doing this eagerly instead of when ZCML runs
-    # because it must be done before ``registerMimeFactories``
-    # is invoked in order to add the mimeType fields if they are missing.
-    # Rewrite so that this can be done as an ZCML action.
-    # _context.action( discriminator=('class_init', tuple(modules)),
-    #                  callable=cls_iio.__class_init__,
-    #                  args=() )
-    cls_iio.__class_init__()
+
+    # Unfortunately, we are doing this eagerly instead of when the
+    # configuration executes its actions runs because it must be done
+    # before ``registerMimeFactories`` is invoked in order to add the
+    # mimeType fields if they are missing. If we deferred it, we would
+    # have to defer registerMimeFactories---and one action cannot
+    # invoke another action or add more actions to the list and still
+    # have any conflicts detected. Using the `order` parameter doesn't help us
+    # much with that, either.
+
+    # The plus side is that now that we are using component_zcml.utility()
+    # to register legacy class factories too, there's not much harm in
+    # initing the class early.
+
+    legacy_factories = cls_iio.__class_init__()
 
     # Now that it's initted, register the factories
     for module in (factory_modules or modules):
         logger.log(loglevels.TRACE,
                    "Examining module %s for mime factories", module)
         registerMimeFactories(_context, module)
+
+    if register_legacy_search_module:
+        for name, factory in _find_factories_in_module(legacy_factories):
+            component_zcml.utility(_context,
+                                   provides=internalization._ILegacySearchModuleFactory,
+                                   component=factory,
+                                   name=name)